adding token validation on all pertinent routes of UserService and Controller

+++ src/main/java/com/app/todos/Services/Users/UserService.java
@@ -2,8 +2,10 @@ package com.app.todos.Services.Users;
 
 import com.app.todos.DTOs.Users.NewUserDTO;
 import com.app.todos.DTOs.Users.UpdateDTO;
+import com.app.todos.Models.Todos.Todo;
 import com.app.todos.Models.Users.User;
 import com.app.todos.Repository.User.UserRepo;
+import com.app.todos.Services.Auth.TokenService;
 import org.springframework.aop.target.LazyInitTargetSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -19,8 +21,12 @@ public class UserService {
     private UserRepo userRepo;
     @Autowired
     private PasswordEncoder passwordEncoder;
+    @Autowired
+    private TokenService tokenService;
 
-    public User get(BigInteger id) {
+    public User get(BigInteger id, String token) {
+        User u = userRepo.findById(id).orElseThrow();
+        if (!tokenService.validate(token).equals(u.getEmail())) return null;
         return userRepo.findById(id).orElseThrow();
     }
 
@@ -30,9 +36,10 @@ public class UserService {
         userRepo.save(u);
     }
 
-    public void update(BigInteger id, UpdateDTO data) {
+    public void update(BigInteger id, UpdateDTO data, String token) {
         User u = userRepo.findById(id).orElseThrow();
 
+        if (!tokenService.validate(token).equals(u.getEmail())) return;
         if (!passwordEncoder.matches(data.currentPassword(), u.getPassword())) return;
 
         String encoded = passwordEncoder.encode(data.newPassword());
@@ -44,7 +51,9 @@ public class UserService {
         userRepo.save(u);
     }
 
-    public void delete(BigInteger id) {
+    public void delete(BigInteger id, String token) {
+        User u = userRepo.findById(id).orElseThrow();
+        if (!tokenService.validate(token).equals(u.getEmail())) return;
         userRepo.deleteById(id);
     }
 }
